# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "User Op"
                ],
                "summary": "User Register",
                "description": "Required => fullname, email, phone, password",
                "operationId": "42744a1bca59abd208368d16de86d85e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "description": "Ad Soyad",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "E-Posta",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "description": "Telefon Numarası",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Parola",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "profile_pic": {
                                        "description": "Profil Resmi",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User Op"
                ],
                "summary": "User Login",
                "description": "Required => email, password",
                "operationId": "dff25add827941151985ecb69773bab5",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "User Op"
                ],
                "summary": "User Profile",
                "description": "Required => token",
                "operationId": "af5f13d9537bca65175b932782d191c8",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/profile/change/pic": {
            "post": {
                "tags": [
                    "User Op"
                ],
                "summary": "User Profile Pic Change",
                "description": "Required => picture",
                "operationId": "2a3b5ce273fd7bd1e141f20a500ece98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "profile_pic": {
                                        "description": "Profil Resmi",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}